[filesystem] Made `watchFileChanges` await
The documentation states that the promise returned will resolve when the
watching will actually be started. Removing the awaiting from there just breaks
that assumption.

A better place to fix the issue is at the consumer end: If you know that the
call might hang, just register a callback and do not wait for it.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>